funk: move global iter API to private header
Discourage use of global iterator, which is inherently incompatible
with concurrent usage.
