+++ src/funk/test_funk_common.hxx
@@ -1,4 +1,4 @@
-#include "fd_funk.h"
+#include "fd_funk_private.h"
 #include <map>
 #include <vector>
 #include <set>
